pyproject.toml
[tool.poetry]
name = "leads-for-brokers-api"
version = "0.1.0"
description = ""
authors = ["Paulo Herrera <pauloherreralopes@gmail.com>"]
readme = "README.md"
packages = [{include = "leads_for_brokers_api"}]

[tool.poetry.dependencies]
python = "3.12.1"
flask = "^3.0.1"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"
psycopg2 = "^2.9.9"
flask-sqlalchemy = "^3.1.1"
flask-restful = "^0.3.10"
sqlalchemy = "^2.0.25"
flask-alembic = "^2.0.1"
flask-httpauth = "^4.8.0"
pyjwt = "^2.8.0"
flask-cors = "^4.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
